your message

+++ routes/api.php
@@ -15,16 +15,19 @@ use App\Http\Controllers\TaskController;
 |
 */
 
-// Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-//     return $request->user();
-// });
+
  Route::post('/signup',[UserController::class, 'signup']);
- Route::get('/sig',[UserController::class, 'getid']);
+
+ Route::middleware('auth:sanctum')->get('/sig',[UserController::class, 'getid']);
+
  Route::middleware('auth:sanctum')->post('/store',[TaskController::class, 'store']);
- Route::patch('/update/{id}',[TaskController::class, 'update']);
- Route::delete('/delete/{id}',[TaskController::class, 'delete']);
+
+ Route::middleware('auth:sanctum')->put('/update',[TaskController::class, 'update']);
+
+
+ Route::get('/getdata/{id}',[TaskController::class, 'getdata']);
+
+ Route::middleware('auth:sanctum')->delete('/delete/{id}',[TaskController::class, 'delete']);
+
  Route::post('/login',[UserController::class, 'login']);
 
-// Route::post('/signup',function (Request $request) {
-//     return response()->json($request);
-// });
